remove old query code and move tests to baseapp test file
